added mapping modules
