+++ MANIFEST.in
@@ -1,4 +1,3 @@
-include AUTHORS.rst
 include CONTRIBUTING.rst
 include LICENSE
 include README.rst
