+++ maria/configs/atmosphere.json
@@ -1,0 +1,57 @@
+{
+    "APEX" : {
+        "description": "",
+        "region" : "chajnantor",
+        "latitude" : -23.0058,
+        "longitude" : -67.7592,
+        "altitude" : 5064,
+        "seasonal": 1,
+        "diurnal": 1,
+        "quantiles": {}
+    },
+
+    "ACT" : {
+        "description": "Atacama Cosmology Telescope",
+        "region" : "chajnantor",
+        "latitude" : -22.958567282386465,
+        "longitude" : -67.7878496702156,
+        "altitude" : 5190,
+        "seasonal": 1,
+        "diurnal": 1,
+        "quantiles": {}
+    },
+
+    "GBT" : {
+        "description": "Green Bank Telescope",
+        "region" : "green_bank",
+        "latitude" : 38.433121, 
+        "longitude" : -79.839835,
+        "altitude" : 818,
+        "seasonal": 1,
+        "diurnal": 1,
+        "quantiles": {}
+    },
+
+    "JCMT" : {
+        "description": "",
+        "region" : "mauna_kea",
+        "latitude" : 19.8228, 
+        "longitude" : -155.477,
+        "altitude" : 4092,
+        "seasonal": 1,
+        "diurnal": 1,
+        "quantiles": {}
+    },
+
+    "SPT" : {
+        "description": "",
+        "region" : "south_pole",
+        "latitude" : 0, 
+        "longitude" : -90,
+        "altitude" : 2800,
+        "seasonal": 1,
+        "diurnal": 1,
+        "quantiles": {}
+    }
+
+}
