+++ maria/configs/mappers.json
@@ -1,0 +1,6 @@
+{
+    "RawBinMapper" : {
+        "resolution": 0.1,
+        "units" : "degrees"
+    }
+}
